requirements.txt
Django==3.2.20
django-cors-headers==4.0.0
Pillow==9.5.0
python-dotenv==1.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2